functions/requirements.txt
azure-functions>=1.18
azure-cosmos>=4.7
azure-storage-blob>=12.19
azure-eventgrid>=4.19
pydantic>=2.0
Pillow>=10.0
psycopg2-binary>=2.9
fiona>=1.9
shapely>=2.0
